Create lib API
