+++ smlibppm/api/__init__.py
@@ -1,0 +1,12 @@
+from smlibppm.core.rect import (
+    draw_rect,
+    draw_rect_rand_colors,
+    fill_rect, 
+    fill_rect_rand_colors,
+) 
+from smlibppm.core.line import (
+    draw_line, 
+    draw_line_rand_colors,
+)
+from smlibppm.core.pixel import put_pixel
+from smlibppm.core.ppm import *
