+++ smlibppm/core/error.py
@@ -1,0 +1,5 @@
+from __future__ import print_function 
+import sys
+
+def eprint(*args, **kwargs):
+    print(*args, file=sys.stderr, **kwargs)
