+++ smlibppm/core/png.py
@@ -1,0 +1,13 @@
+HEADER = '137 80 78 71 13 10 26 10'
+def parse_png_header(fd: str) -> bool:
+    """[parse a png header]
+
+    Args:
+        fd (str): [path to png]
+    """
+
+
+def test():
+    
+
+test()
